+++ src/main/java/net/geforcemods/securitycraft/network/ServerProxy.java
@@ -1,19 +1,3 @@
 package net.geforcemods.securitycraft.network;
 
-import net.minecraft.entity.player.EntityPlayer;
-
-public class ServerProxy implements IProxy {
-	@Override
-	public void registerEntityRenderingHandlers() {}
-
-	@Override
-	public void registerRenderThings() {}
-
-	@Override
-	public void registerVariants() {}
-
-	@Override
-	public EntityPlayer getClientPlayer() {
-		return null;
-	}
-}
+public class ServerProxy implements IProxy {}