+++ Changelog.md
@@ -3,10 +3,14 @@
 - New: Support for teams from FTB Utilities when having team ownership enabled in the config
 - New: Laser blocks can now be configured to have lasers on specific sides be disabled by using a Smart Module
 - New: Configs "sentryAttackableEntitiesAllowlist" and "sentryAttackableEntitiesDenylist" for controlling which entities a Sentry can and cannot attack
+- New: The Alarm can now play any sound when a Smart Module is installed
+- New: The Alarm can now be disabled with the Universal Block Modifier
+- New: Alarm option "Reset Cooldown" to set whether the cooldown between alarm sounds should reset once the alarm is powered on
 - Change: Lasers and inventory scanner fields are now removed silently
 - Change: Inventory scanners and password-protected chests now only drop their modules if the last scanner/chest is removed
 - Change: Changing the owner of a Reinforced Hopper now needs validation of the new owner (akin to reinforced pistons)
 - Change: New look for the Claymore
+- Change: The Alarm's "Delay" option can now be found in the new GUI of the Alarm, accessible by rightclicking it
 - API: IModuleInventory#shouldDropModules to determine whether a block should drop its modules when broken
 - API: Numerous renames across the whole codebase to make porting from newer versions easier going forward
 - Fix: Turning off camera rotation or the Sonic Security System leads to (extreme) stuttering of the camera/SSS